debian: Check for accurate Debian changelog version at build time too.
When we increment the Open vSwitch version number, we tend to forget to
update it in debian/changelog at the same time.  Right now this gets
fixed up automatically at "make dist" time, but it's even better if we can
always have it be correct in the repository.  This commit should help with
that, by making both "make" and "make dist" refuse to proceed if the
version number is out of sync.