ofproto: Add Nicira extension to OpenFlow multiple flow tables support.
Some switching hardware has multiple flow tables with varying capabilities.
For example, one flow table may be able to match only on L2 fields and
another only on L3 fields, and the tables may support different sets of
actions.  Tables may also have overlapping capabilities.  Until now, there
has been no way for OpenFlow controllers to specify a particular table to
put a flow into, which is necessary in the case of overlapping
capabilities.  OpenFlow 1.1 will standardize a set of features to handle
this, but for use until then this commit adds a Nicira extension with
simple support for explicitly manipulating specific tables.

In short, this commit uses the upper 8 bits of the 'command' field in
ofp_flow_mod to specify a table.  The details are described in
include/openflow/nicira-ext.h.

I've confirmed that this doesn't break anything in my simple test cases
when the extension is not used.  I haven't yet tested the extension itself.