vswitch: Clarify and expand purpose of external_ids columns.
Until now we have intended external_ids for use by external frameworks, but
only for storing "identifiers" such as UUIDs and names.  But there's no
reason that frameworks can't use it for whatever purposes they like.  This
commit rewords the text describing the external_ids columns to be more
expansive.

CC: Peter Balland <peter@nicira.com>