ofproto: Improve OFPT_FLOW_MOD error reporting.
Error reporting to the controller for "flow_mod" operations was poor
because WDP was reporting errors as Unix errno values and nothing was
translating them into OpenFlow errors to send to the controller.  This
commit fixes the problem by making the functions in question instead use
OpenFlow error numbers and, in cases where there is no appropriate number,
using a vendor extension error number.