REPORTING-BUGS: Rewrite based on experience.
Burying the description of the problem, which is usually the most important
part in my experience, at the bottom of the REPORTING-BUGS file might be
the reason why we don't get good descriptions sometimes.  It is certainly
not the reason in other cases, but we might as well prioritize a bit
better.