debian: Add procps to Build-Depends.
"make check" runs "ps" so the procps package is necessary.  With this
change, I was able on Friday to make the Open vSwitch Debian packages build
successfully under the "sbuild" suite used by the Debian autobuilders, so
I hope that this fixes at least some of the Debian buildd problems.

CC: Simon Horman <horms@verge.net.au>
Debian bug #602891.