flow: Delete unused FWW_VLAN_TCI bit.
This wasn't used intentionally anywhere, but some code was turning it on
accidentally (because it was part of FWW_ALL) and other code was not, which
caused confusion.  In particular, the NXM code turned it on by default
and the OpenFlow 1.0 code did not, which caused flow stat requests to
return different results depending on format.  Deleting it fixes the bug.

Reported-by: Natasha Gude <natasha@nicira.com>
CC: Natasha Gude <natasha@nicira.com>