capwap: Bind address should be big endian.
CAPWAP creates a UDP socket that accepts packets from any address using
INADDR_ANY.  IP addresses should be in network byte order but that
constant is in host byte order, so use htonl.  However, this is not a
real bug since the value of INADDR_ANY is 0.

Found with sparse.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Acked-by: Ben Pfaff <blp@nicira.com>