brcompat: Simplify generation of bridge ID.
Currently we use a fairly complicated method of generating the
bridge ID, since the actual struct is only available in a header
file private to the Linux bridge.  The current method appears to
be correct but is difficult to reason about.  This replaces it
with a simple memcpy, which is more analogous to what the Linux
bridge does.

Flagged by sparse.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Acked-by: Ben Pfaff <blp@nicira.com>