ofproto: Change xlate_actions() to take a structure.
An upcoming commit has a need to give xlate_actions() another parameter,
but it already has too many.  This commit improves the situation by making
xlate_actions()'s caller fill in a structure instead.

The action_xlate_ctx structure is kind of big and unwieldy because it
include a struct odp_actions, which is about 16 kB.  But work underway will
change that to a "struct ofpbuf", which is much more reasonable.