flow: Make size of flow struct a multiple of 8.
The compiler wants to pad structures to a multiple of the native
datatype for the architecture, so a multiple of 4 on 32-bit platforms
and a multiple of 8 on 64-bit.  Currently the size struct flow is
a multiple of 4, so the total size with padding varies depending on
the architecture, causing build asserts to fail.  This explicitly pads
it out to a multiple of 8 for consistency.