datapath-protocol: Include netlink.h.
On older kernels that don't have if_link.h, we use our own, limited
version.  This version doesn't include the netlink header, causing
problems where we were relying on it to define the types in
datapath-protocol.h.  Therefore, directly include it, since it is
better to be explicit about it anyways.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Acked-by: Ben Pfaff <blp@nicira.com>