debian: Modify ovs-bugtool to include collectd rrd logs
collectd is a utility that stores rrd logs in the directory
/var/lib/collectd/rrd/ by default. rrd logs are created when
collectd is configured to output logs in Round Robin Database
format.

With this change, ovs-bugtool will include collectd logs in the
output bundle from /var/lib/collectd/rrd.  The output bundle will
include all sub-directories of the specified collectd directory.

ovs-bugtool checks if a directory exists before it attempts to
include the directory contents in the output bundle. So, running
ovs-bugtool on a system that does not have collectd logs should
not negatively affect the behavior of ovs-bugtool.